sidecar/weather-api/requirements.txt
flask>=2.3.0
flask-cors>=4.0.0
PyJWT>=2.8.0
requests>=2.31.0
